Refactor header component to use new file structure
