+++ app/_components/prompt-box.tsx
