+++ app/page.tsx
@@ -1,8 +1,8 @@
 import { Badge } from "@/components/ui/badge";
-import GenerateBox from "./components/prompt-box";
-import IntroSection from "./components/intro-section";
-import AppWrapper from "./components/react-query-provider";
-import PromptBox from "./components/prompt-box";
+import GenerateBox from "./_components/prompt-box";
+import IntroSection from "./_components/intro-section";
+import AppWrapper from "./_components/react-query-provider";
+import PromptBox from "./_components/prompt-box";
 
 export default function Home() {
   return (
